VERY experimental commit:
shared SelectFeature control for WFS and geometryLayer data

getting rid of ooi_click endpoint. Instead, a 'clicked' endpoint transmits lon-lat data including either an 'ooi' or 'area' property depending on whether something was hit on the geometry or WFS layer. If neither is present, nothing was clicked but a blank spot on the map.